feat: Add global panic hook
