+++ common/logger/Cargo.toml
@@ -11,4 +11,3 @@ repository = "https://github.com/nervosnetwork/muta"
 log = "0.4"
 log4rs = "0.8"
 json = "0.12"
-backtrace = "0.3"
